@@ -8,22 +8,15 @@
     "gatsby"
   ],
   "scripts": {
-    "develop": "gatsby develop",
-    "start": "gatsby develop",
-    "build": "gatsby build",
-    "serve": "gatsby serve",
-    "clean": "gatsby clean"
+    "start": "parcel index.html"
   },
   "dependencies": {
     "@uiw/react-md-editor": "^2.1.8",
-    "babel-plugin-styled-components": "^1.12.0",
     "bootstrap": "^4.6.0",
-    "gatsby": "^2.32.3",
-    "gatsby-plugin-react-helmet": "^3.10.0",
-    "gatsby-plugin-styled-components": "^3.10.0",
-    "react": "^16.13.1",
+    "parcel-bundler": "^1.12.4",
+    "react": "^16.14.0",
     "react-bootstrap": "^1.4.3",
-    "react-dom": "^16.13.1",
+    "react-dom": "^16.14.0",
     "react-helmet": "^6.1.0",
     "styled-components": "^5.2.1"
   }
